Add sorting of scripts in export
For the export to referencedata_scripts.xml, script groups and
scripts are sorted on

* category, i.e. group vs script (column ScriptCategory)
* GUID (column ScriptName)

This is to have a more consistent output from export to export and
thus make it easier to compare referencedata_scripts.xml across
branches.

For the export to README.md, script groups are sorted on name, and
scripts in a script group are sorted on name as well.

This is to have  README.md file that matches the alphabetic sorting
in EA.