Update script export for use with QEA files and EA 16
Differences from EAPX and EA 15:

- The results of sql queries have an encoding declaration with value
"UTF-16" (as opposed to no encoding declaration, implying UTF-8). The
code is updated to handle different encoding declarations and a
possible byte order mark in the beginning of an sql query result.
- The newlines from the scripts themselves are with EA 16 present in
the result of the sql query used to retrieve the scripts and the script
groups. In order to deal with this change:
  - the newlines are now removed from the reference data XML file, to
match the output that EA gives when exporting scripts via "Export
Reference Data";
  - the newlines are not added anymore when creating and saving the
scripts as separate files.