Update how the name and the version of a logical data model are set
Many models do have their tagged values set correctly, so we cannot
rely on that when exporting models. Therefore, use the UML package name
as the name of the model, and check whether perhaps the version is set
on the package if it is not set in a tagged value.